feat: add MVU commit linter CLI
